removed a compilation warning

+++ src/noc/routing/slb/slb-routing.cc
@@ -216,7 +216,7 @@ namespace ns3
       }
 
     UniformVariable randomVariable;
-    device = selectedDevices[randomVariable.GetValue (0, selectedDevices.size ())];
+    device = selectedDevices[(int)randomVariable.GetValue (0, selectedDevices.size ())];
 
     NS_LOG_DEBUG ("The net device " << device->GetAddress () << " was selected for routing");
